destructuring
Деструктуризация кода в эксель версии
